Factor out listener set object from the integrated dispatcher.
